+++ CHANGES.md
@@ -9,6 +9,7 @@ Change Log
 * `quantizedAttributes` has an optional `normalized` flag to use the glTF 1.0.1 `accessor.normalized` for a higher precision decode matrix.
 * Fixed an issue where pipeline extras are not removed when running `Pipeline.processJSON` and `Pipeline.processFile`.
 * Added support for generating hard normals with the `-f` flag and for removing normals with `-r`. [#173](https://github.com/AnalyticalGraphicsInc/gltf-pipeline/pull/173)
+* The `_3DTILESDIFFUSE` semantic is added to the model's technique when `optimizeForCesium` is true.
 
 ### 0.1.0-alpha4 - 2016-08-25